pyproject.toml
[tool.poetry]
name = "aegis-pentest-automation"
version = "1.0.0"
description = "Automatización de pentesting web con GUI Qt - Wrapper de herramientas de seguridad"
authors = ["Aegis Security <aegis@security.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "aegis_pentest"}]

[tool.poetry.dependencies]
python = "^3.9"
python-nmap = "^0.7.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
rich = "^13.7.0"
inquirer = "^3.2.0"
tabulate = "^0.9.0"
prompt-toolkit = "^3.0.43"
beautifulsoup4 = "^4.12.2"
colorama = "^0.4.6"
python-docx = "^1.1.0"
jinja2 = "^3.1.2"
aiohttp = "^3.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-html = "^4.1.0"
pytest-xdist = "^3.5.0"
pytest-mock = "^3.12.0"
psutil = "^5.9.0"
coverage = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aegis = "aegis_cli:main"

# Comando único para ejecutar todos los tests
test = "tests.scripts.run_all_tests:main"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg-info"
] 